refactor: move app controller and styles to app layer
